src/agent_c_core/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "agent_c-core"
version = "0.1.3"
dependencies = [
    "pydantic==2.9.2",
    "openai==1.86.0",
    "tiktoken==0.9.0",
    "boto3<1.38.28,>=1.38.2",
    "anthropic[bedrock]==0.54.0",
    "diskcache==5.6.3",
    "observable==1.0.3",
    "sounddevice>=0.5.2",
    "numpy>=2.3.0",
    "soundfile",
    "pyyaml",
    "mcp>=1.6.0",
    "structlog>=25.3.0",
    "Pygments>=2.19.1",
    "pymdown-extensions>=10.15",
    "uszipcode>=1.0.1",
    "pgeocode>=0.5.0",
    "timezonefinder>=6.5.9",
    "jinja2>=3.1.6",
    "transitions>=0.9.2",
]
requires-python = ">=3.10"
authors = [
  {name = "Donavan Stanley", email = "donavan.stanley@centricconsulting.com"},
  {name = "Joe Ours", email = "joesph.ours@centricconsulting.com"},
]
maintainers = [
  {name = "Donavan Stanley", email = "donavan.stanley@centricconsulting.com"},
]
description = "Core functionality for Agent C"
readme = "README.MD"
license = { text = "Business Source License 1.1" }
keywords = ["gpt", "claude", "agents", "ai", "centric"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

[project.urls]
Homepage = "https://github.com/centricconsulting/agent_c_framework"

